# doc/api.yaml
openapi: 3.0.0
info:
  title: BTD Project
  description: Sample NodeJS BTD spec
  version: '1.0.0'
servers:
  - url: http://localhost:3000
    description: dev server
security:
  - BearerAuth: []
...
paths:
  /api/me:
    get:
      tags:
          - me
      summary: Get current user
      security:
        - BearerAuth: []
      description: Get current user
      responses:
        '200':
          description: Response information depends on role
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/Users'
                  - $ref: '#/components/schemas/RestrictedUser'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Server Error
          content:
            'text/html':
              schema:
                $ref: '#/components/schemas/ServerError'
    put:
      tags:
        - me
      summary: Update current user
      security:
        - BearerAuth: []
      description: Update current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Server Error
          content:
            'text/html':
              schema:
                $ref: '#/components/schemas/ServerError'
    
    delete:
      tags:
        - me
      summary: Delete current user
      security:
        - BearerAuth: []
      description: Delete current user
      responses:
        '200':
          description: Me deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Server Error
          content:
            'text/html':
              schema:
                $ref: '#/components/schemas/ServerError'
...
  /api/users/{userId}:
    get:
      tags:
        - user
      summary: Get user by ID
      security:
        - BearerAuth: []
      description: Get user by ID
      parameters:
       - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Server Error
          content:
            'text/html':
              schema:
                $ref: '#/components/schemas/ServerError'
                
    put:
      tags:
        - user
      summary: Update user
      security:
        - BearerAuth: []
      description: Update user
      parameters:
       - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Server Error
          content:
            'text/html':
              schema:
                $ref: '#/components/schemas/ServerError'
    
    delete:
      tags:
        - user
      summary: Delete user
      security:
        - BearerAuth: []
      description: Delete user
      parameters:
       - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: User deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Server Error
          content:
            'text/html':
              schema:
                $ref: '#/components/schemas/ServerError'
  /api/products:
    get:
      tags:
        - product
      summary: Get all products
      security:
        - BearerAuth: []
      description: Get all products
      responses:
        '200':
          description: Gets list of products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Products'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Server Error
          content:
            'text/html':
              schema:
                $ref: '#/components/schemas/ServerError'
    post:
      tags:
        - product
      summary: Create product
      security:
        - BearerAuth: []
      description: Create new product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Server Error
          content:
            'text/html':
              schema:
                $ref: '#/components/schemas/ServerError'
  /api/products/{productId}:
    get:
      tags:
        - product
      summary: Get product by ID
      security:
        - BearerAuth: []
      description: Get product by ID
      parameters:
       - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/NotFound'
        default:
          description: Unexpected error
          
    put:
      tags:
        - product
      summary: Update product by ID
      security:
        - BearerAuth: []
      description: Update product by ID
      parameters:
       - $ref: '#/components/parameters/ProductId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: Product updated
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/NotFound'
        default:
          description: Unexpected error   
    
    delete:
      tags:
        - product
      summary: Delete product by ID
      security:
        - BearerAuth: []
      description: Delete product by ID
      parameters:
       - $ref: '#/components/parameters/ProductId'
      responses:
        '200':
          description: Product deleted
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/NotFound'
        default:
          description: Unexpected error  
  /api/categories:
    get:
      tags:
        - category
      summary: Get all categories 
      security:
        - BearerAuth: []
      description: Get all categories
      responses:
        '200':
          description: Gets list of categories
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Categories'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Server Error
          content:
            'text/html':
              schema:
                $ref: '#/components/schemas/ServerError'
                
    post:
      tags:
        - category
      summary: Create new category
      security:
        - BearerAuth: []
      description: Create new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BodyCategory'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Server Error
          content:
            'text/html':
              schema:
                $ref: '#/components/schemas/ServerError'
  /api/categories/{categoryId}:
    get:
      tags:
        - category
      summary: Get category by ID
      security:
        - BearerAuth: []
      description: Get category by ID
      parameters:
       - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Category response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/responses/NotFound'
        default:
          description: Unexpected error
          
    put:
      tags:
        - category
      summary: Update category
      security:
        - BearerAuth: []
      description: Update category
      parameters:
       - $ref: '#/components/parameters/CategoryId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '200':
          description: Category updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Server Error
          content:
            'text/html':
              schema:
                $ref: '#/components/schemas/ServerError'
    delete:
      tags:
        - category
      summary: Delete category
      security:
        - BearerAuth: []
      description: Delete category
      parameters:
       - $ref: '#/components/parameters/CategoryId'
      responses:
        '200':
          description: Category deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        default:
          description: Server Error
          content:
            'text/html':
              schema:
                $ref: '#/components/schemas/ServerError'
...
components:
  schemas:
...
    BodyCategory:
      type: object
      properties:
        name:
          type: string
          example: Category name
        slug:
          type: string
          example: categoryslug
    Users:
      type: array
      items:
        $ref: "#/components/schemas/User" 
    User: 
      type: object
      required:
        - name
        - role
      properties:
        id:
          type: number
          example: 1
        name:
          type: string
          example: Jonh Doe
        role:
          type: string
          enum:
            - Admin
            - User
          example: User
        isActive:
          type: boolean
          example: false
        email:
          type: string
          example: jonh-doe@example.com
        password:
          type: string
          example: '123456'
    RestrictedUser:
      type: object
      required:
        - name
        - role
      properties:
        id:
          type: number
          example: 1
        name:
          type: string
          example: Jonh Doe
        email:
          type: string
          example: jonh-doe@example.com
        password:
          type: string
          example: '123456'
...
    Products:
      type: array
      items:
        $ref: "#/components/schemas/Product"
    Product: 
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 4
        name:
          type: string
          example: Product Name
        slug:
          type: string
          example: ThisProductJonhDoe
        created date:
          type: string
          format: date
          example: '2021-03-20'
        number of views:
          type: integer
          format: int64
          example: 14
        price:
          type: number
          example: 400
        description:
          type: string
          example: About JonhDoeJonhDoe JonhDoeJonhDoe
        images:
          type: string
          format: binary
    Categories:
      type: array
      items:
        $ref: "#/components/schemas/Category"
    Category: 
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 4
        name:
          type: string
          example: Category Name
        slug:
          type: string
          example: ThisCategorytJonhDoe
...
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
      required:
        - code
        - message
    ServerError:
      type: object
      properties:
        status: 
          type: integer
          enum:
           - 500
        error:
          type: string
          example: 'internal-server-error'
        message:
          type: string
          example: 'Something went wrong'
        extraDetails:
          type: string
          example: 'Extra error details'
  responses:
...
    UnauthorizedError:
      description: Access token is missing or invalid
    NotFound:
      description: The specified resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  parameters:
...
    UserId:
      name: userId
      in: path
      description: ID of user to use
      required: true
      schema:
        type: number
        example: 1
    ProductId:
      name: productId
      in: path
      description: ID of product to use
      required: true
      schema:
        type: number
        example: 1
    CategoryId:
      name: categoryId
      in: path
      description: ID of category to use
      required: true
      schema:
        type: number
        example: 1
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT